refactor: move tinyagent stream loop into agent_components
Made-with: Cursor
